Separate service-accounts list
